Incorporate rxjs stream into old work
