+++ src/index.ts
@@ -10,8 +10,6 @@ export * from "./realtime";
 export * from "./combo";
 export * from "./types";
 
-export { RxSlpStream } from "./utils/rxSlpStream";
-export { RxSlpRealTime } from "./realtime/rxRealTime";
 
 // Re-export SlippiGame for convenience
 export { SlippiGame } from "slp-parser-js";
