+++ test/slpStream.spec.ts
@@ -2,8 +2,18 @@ import sinon from "sinon";
 
 import { SlpStream } from "../src";
 import { pipeFileContents } from "./helpers";
+import { Subscription } from "rxjs";
 
 describe("SlpStream", () => {
+  let subscriptions: Array<Subscription>;
+
+  beforeAll(() => {
+    subscriptions = [];
+  });
+
+  afterAll(() => {
+    subscriptions.forEach(s => s.unsubscribe());
+  });
 
   describe("when reading from a standard slp file", () => {
     it("emits exactly one game start and one game end event", async () => {
@@ -11,8 +21,9 @@ describe("SlpStream", () => {
       const gameEndSpy = sinon.spy();
 
       const slpStream = new SlpStream({ singleGameMode: true });
-      slpStream.on("gameStart", gameStartSpy);
-      slpStream.on("gameEnd", gameEndSpy);
+      const unsubGameStart = slpStream.gameStart$.subscribe(gameStartSpy);
+      const unsubGameEnd = slpStream.gameEnd$.subscribe(gameEndSpy);
+      subscriptions.push(unsubGameStart, unsubGameEnd);
 
       await pipeFileContents("slp/Game_20190810T162904.slp", slpStream);
